Added LICENSE file.  I would be happy to see all the GPL coments disappear from single source files in the future.
Added a Boundable interface to the models package.  This interface is thought
to be a tool for estimating the target interval given a source interval for a
CoordinateTransform that is a continuous function or can be approximated as a
continuous function.  In other words: find the bounding box for a given source
box.