Added experimental ExtendedPluginFilter for statistic outlier removal using Integral images.
The plugin calculates mean and standard deviation of a block
at each pixel and, if the pixel's intensity is smaller or larger
than mean+-std*k fills it with neighboring pixel content.

Not yet implemented for anything but FloatProcessor which
stems only from limited time.

TODO Implement for integer types and RGB.