Contributed a solution to a long standing problem that the global optimization might not converge for some transformation models like Homography if it is blindly initialized.
Now there is a function TileConfiguration.preAlign() that estimates the final configuration
by propagating pairwise alignments through the TileConfiguration. It takes into account if
there is more than one fixed tile.

It offers two advantages:
a) it is very likely to converge even for "crazy" models as it is quite close to the final solution
b) it converges much much quicker, which might be interesting for large datasets