Fix versions of the modules
I missed that in the previous round: you still cannot define a
project's version by properties. The idea is that the version is
stable, and that precludes the version from being volatile (especially
when there is only a -SNAPSHOT binding to the parent which defines
that property).

However, you *can* inherit the version from the <parent> block. So
let's do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>