POM: use pom-scijava for repeatable builds
Using a SNAPSHOT parent can cause the parent configuration to shift out
from under the project, resulting in an irreproducible build. Let's use
a release version of pom-scijava instead, to improve build stability.

We use pom-scijava 1.125 because it is the most recent version to
successfully imbue the needed "use=false" hack to maven-javadoc-plugin.
Without that hack, an NPE is thrown during javadoc generation because
the javadoc tool has a long-standing bug when generating package usage
information for classes in the default package.