bump to development cycle 1.0.0-SNAPSHOT
transforms and models operate on double precision instead of float

This is a significant break in the API.  I deliberately have not
preserved previous behavior in deprecated methods because full
compatibility was impossible to guarantee (i.e. access to internal
float[] vectors is incompatible with double[]) and breaking everything
clean makes it easier to adopt.