New method BlockMatching.matchByMaximalPMCCFromPreScaledImages (#49)
* New method BlockMatching.matchByMaximalPMCCFromPreScaledImages
which does what is says: it assumes the source and target FloatProcessor
instances provided as arguments have already been scaled, normalized
and masked. In addition, it removes the need for an approximate
transformation, assuming that either there isn't a large one (meaning
images are largely already in register) or that the caller took care
of that.

This new method is a performance optimization over the full-fledged
matchByMaximalPMCC method that takes the full sized images, their masks
and the approximate transform. By avoiding operations that the caller
knows are not necessary, lots of work is saved when e.g. comparing
one section to the adjacent, the second adjacent, the third adjacent, etc.,
avoiding lots of recomputations (Gaussian blur, scaling, contrast
normalization and masking out with NaNs).

* The ErrorStatistic is not needed for the uses of matchByMaximalPMCCFromPreScaledImages.

* The image is already scaled, so the translation has to be scaled too.